implement UndoRedoManager
