+++ packages/collaboration-manager/src/CollaborationManager.ts
@@ -2,6 +2,7 @@ import type { EditorJSModel, ModelEvents } from '@editorjs/model';
 import { EventType, TextAddedEvent, TextRemovedEvent } from '@editorjs/model';
 import { Operation, OperationType } from './Operation.js';
 import { Transformer } from './Transformer.js';
+import { UndoRedoManager } from './UndoRedoManager.js';
 
 /**
  * CollaborationManager listens to EditorJSModel events and applies operations
@@ -13,18 +14,12 @@ export class CollaborationManager {
   #model: EditorJSModel;
 
   /**
-   * Local stack of operations to undo
+   * UndoRedoManager instance to manage undo/redo operations
    */
-  #undoStack: Operation[] = [];
+  #undoRedoManager: UndoRedoManager;
 
   /**
-   * Local stack of operations to redo
-   */
-  #redoStack: Operation[] = [];
-
-
-  /**
-   * Flag to control whether events should be handled
+   * Flag to control whether events should be handled to avoid putting operations to the stack on undo/redo
    */
   #shouldHandleEvents = true;
 
@@ -36,6 +31,7 @@ export class CollaborationManager {
    */
   constructor(model: EditorJSModel) {
     this.#model = model;
+    this.#undoRedoManager = new UndoRedoManager();
     model.addEventListener(EventType.Changed, this.#handleEvent.bind(this));
   }
 
@@ -43,29 +39,7 @@ export class CollaborationManager {
    * Undo last operation in the local stack
    */
   public undo(): void {
-    const operation = this.#undoStack.pop();
-
-    if (operation === undefined) {
-      return;
-    }
-
-    // Disable event handling
-    this.#shouldHandleEvents = false;
-
-    const inversedOperation = Transformer.inverse(operation);
-
-    this.applyOperation(inversedOperation);
-    this.#redoStack.push(operation);
-
-    // Re-enable event handling
-    this.#shouldHandleEvents = true;
-  }
-
-  /**
-   * Redo last undone operation in the local stack
-   */
-  public redo(): void {
-    const operation = this.#redoStack.pop();
+    const operation = this.#undoRedoManager.undo();
 
     if (operation === undefined) {
       return;
@@ -75,7 +49,27 @@ export class CollaborationManager {
     this.#shouldHandleEvents = false;
 
     this.applyOperation(operation);
-    this.#undoStack.push(operation);
+    this.#undoRedoManager.putToRedoStack(Transformer.inverse(operation));
+
+    // Re-enable event handling
+    this.#shouldHandleEvents = true;
+  }
+
+  /**
+   * Redo last undone operation in the local stack
+   */
+  public redo(): void {
+    const operation = this.#undoRedoManager.redo();
+
+    if (operation === undefined) {
+      return;
+    }
+
+    // Disable event handling
+    this.#shouldHandleEvents = false;
+
+    this.applyOperation(operation);
+    this.#undoRedoManager.putToUndoStack(operation);
 
     // Re-enable event handling
     this.#shouldHandleEvents = true;
@@ -138,8 +132,8 @@ export class CollaborationManager {
     }
 
     if (operation !== null) {
-      this.#undoStack.push(operation);
-      this.#redoStack = [];
+      this.#undoRedoManager.putToUndoStack(operation);
+      this.#undoRedoManager.flushRedoStack();
     }
   }
 }
